Connect app logs and event logs via common metrics
(+) Add partitionId to TaskInfo
(+) Add TaskContext class to Pyspark. The code is taken from a pull
    request by Holden Karau: https://github.com/apache/spark/pull/16211

As a result, we get Stage ID, Partition ID and attempt number in task
metrics and we can also log the same info to Pyspark's stderr. This
gives us an opportunity to associate log records in stderr and metrics
info.